Yellow Squares
Yellow squares do 1 of 3 things: destroy, move, or place a block. If
destroying or moving a block, it first moves to said block. When
destroying it, it deletes the block and itself however when moving it,
the block is moved offscreen and goes to place it.
When placing or the second phase of moving it. it moves to a new
location and creates or moves the block onscreen and deletes itself.
Yellow squares spawn every minute after 90 seconds at the location of
the player.